+++ test/runtime/samples/dynamic-element-event-handler2/main.svelte
@@ -1,0 +1,6 @@
+<script>
+  export let tag;
+  export let handler;
+</script>
+
+<svelte:element this={tag} on:click={handler}>Foo</svelte:element>