+++ test/runtime/samples/dynamic-element-variable/main.svelte
@@ -1,0 +1,6 @@
+<script>
+  export let tag = "div";
+  export let text = "Foo";
+</script>
+
+<svelte:element this={tag}>{text}</svelte:element>